StudyPress 1.1 release
git-svn-id: https://plugins.svn.wordpress.org/studypress/trunk@1151491 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ Class/CourseManager.class.php
@@ -15,6 +15,7 @@ class CourseManager
 
     }
 
+    
 
 
     public function add(Course $course)
@@ -23,11 +24,12 @@ class CourseManager
 
             StudyPressDB::COL_NAME_COURSE => $course->getName(),
             StudyPressDB::COL_DESCRIPTION_COURSE => $course->getDescription(),
+            StudyPressDB::COL_PICTURE_COURSE => $course->getPictureId(),
         );
 
        $this->_access->insert(StudyPressDB::getTableNameCourse(), $a);
 
-
+        
         $idCourse = $this->_access->getLastInsertId();
 
 
@@ -35,11 +37,13 @@ class CourseManager
 
 
 
-
         $this->addCategories($idCourse,$course->getCategories());
 
 
 
+        $this->addGroupsBP($idCourse,$course->getGroupsBP());
+
+        
         $this->addUsers($idCourse,$course->getAuthors());
 
 
@@ -57,8 +61,8 @@ class CourseManager
     public function deleteAllCategories($id){
 
         $this->_access->delete(
-            StudyPressDB::getTableName_CourseCategory(),
-            array(StudyPressDB::COL_ID_COURSE_CAT_N_COURSE => $id)
+            StudyPressDB::getTableName_CourseCategory(), 
+            array(StudyPressDB::COL_ID_COURSE_CAT_N_COURSE => $id)  
         );
 
     }
@@ -73,6 +77,30 @@ class CourseManager
             );
 
             $this->_access->insert(StudyPressDB::getTableName_CourseCategory(), $a);
+        }
+    }
+
+
+    public function deleteAllGroupsBP($id){
+
+        $this->_access->delete(
+            StudyPressDB::getTableName_GroupCourse(), 
+            array(StudyPressDB::COL_ID_COURSE_GROUP => $id)  
+        );
+
+    }
+
+
+    public function addGroupsBP($courseId,array $groups)
+    {
+        foreach ($groups as $g) {
+            $a = array(
+
+                StudyPressDB::COL_ID_GROUP_BP => $g,
+                StudyPressDB::COL_ID_COURSE_GROUP => $courseId,
+            );
+
+            $this->_access->insert(StudyPressDB::getTableName_GroupCourse(), $a);
         }
     }
 
@@ -100,11 +128,30 @@ class CourseManager
     }
 
 
+
+    public function getGroupsBPId($courseId)
+    {
+        $courseId = (int) $courseId;
+
+        $result = $this->_access->getResults($this->_access->prepare(
+                "SELECT " . StudyPressDB::COL_ID_GROUP_BP ." FROM " . StudyPressDB::getTableName_GroupCourse() ." WHERE " . StudyPressDB::COL_ID_COURSE_GROUP . " = '%d'",$courseId)
+        );
+
+        $ids = array();
+        foreach ($result as $row) {
+            $ids[] = $row[StudyPressDB::COL_ID_GROUP_BP];
+        }
+
+
+        return $ids;
+    }
+
+
     public function deleteAllUsers($id){
 
         $this->_access->delete(
-            StudyPressDB::getTableName_CourseUsers(),
-            array(StudyPressDB::COL_ID_COURSE_CAT_N_COURSE => $id)
+            StudyPressDB::getTableName_CourseUsers(), 
+            array(StudyPressDB::COL_ID_COURSE_CAT_N_COURSE => $id)  
         );
 
     }
@@ -145,27 +192,33 @@ class CourseManager
 
 
 
-
+    
     public function update($id, Course $course)
     {
         $this->_access->update(
-            StudyPressDB::getTableNameCourse(),
+            StudyPressDB::getTableNameCourse(), 
             array(
 
                 StudyPressDB::COL_NAME_COURSE => $course->getName(),
                 StudyPressDB::COL_AVANCEMENT_COURSE => $course->getAvancement(),
+                StudyPressDB::COL_PICTURE_COURSE => $course->getPictureId(),
                 StudyPressDB::COL_ID_POST_COURSE => $course->getPostId(),
-                StudyPressDB::COL_DESCRIPTION_COURSE => $course->getDescription(),
+                StudyPressDB::COL_DESCRIPTION_COURSE => $course->getDescription(),  
             ),
-            array(StudyPressDB::COL_ID_COURSE => $id)
+            array(StudyPressDB::COL_ID_COURSE => $id)  
         );
 
 
-
+        
         $this->deleteAllCategories($id);
 
         $this->addCategories($id,$course->getCategories());
 
+
+        
+        $this->deleteAllGroupsBP($id);
+
+        $this->addGroupsBP($id,$course->getGroupsBP());
 
 
         $this->deleteAllUsers($id);
@@ -173,9 +226,7 @@ class CourseManager
         $this->addUsers($id,$course->getAuthors());
 
 
-
         $childrens = PostWP::getChildrenPost($course->getPostId());
-
 
 
 
@@ -189,34 +240,44 @@ class CourseManager
         }
 
 
+        $this->updatePost($course);
+
+
     }
 
-
+    
     public function delete($id)
     {
         $id = (int)$id;
 
         $course = $this->getById($id);
 
+        if($course)
+        {
+           
+            $this->deleteAllCategories($id);
 
-        $this->deleteAllCategories($id);
+
+            $this->deleteAllUsers($id);
 
 
-        $this->deleteAllUsers($id);
+            $this->unpost($course);
 
 
-        $this->unpost($course);
+            
 
+
+        }
 
 
 
         $this->_access->delete(
-            StudyPressDB::getTableNameCourse(),
-            array(StudyPressDB::COL_ID_COURSE => $id)
+            StudyPressDB::getTableNameCourse(), 
+            array(StudyPressDB::COL_ID_COURSE => $id)  
         );
     }
 
-
+   
     public function isError()
     {
         return ($this->_access->getLastError() == '') ? false : true;
@@ -227,12 +288,12 @@ class CourseManager
         return $this->_access->getLastError();
     }
 
-
+    
     public function getAll()
     {
 
         $courses = array();
-        $result = $this->_access->getResults("SELECT * FROM " .  StudyPressDB::getTableNameCourse());
+        $result = $this->_access->getResults("SELECT * FROM " .  StudyPressDB::getTableNameCourse() ." ORDER BY " . StudyPressDB::COL_ID_COURSE ." DESC ");
 
         foreach ($result as $row) {
 
@@ -245,6 +306,7 @@ class CourseManager
 
             $course->setCategories($this->getCategoriesId($course->getId()));
 
+            $course->setGroupsBP($this->getGroupsBPId($course->getId()));
 
 
             $course->setAuthors($this->getUsersId($course->getId()));
@@ -272,8 +334,6 @@ class CourseManager
 
 
 
-
-
     public static function returnedCourse($row)
     {
         return (
@@ -283,13 +343,14 @@ class CourseManager
                 'id'          => (int) $row[StudyPressDB::COL_ID_COURSE],
                 'name'        =>       $row[StudyPressDB::COL_NAME_COURSE],
                 'description' =>       $row[StudyPressDB::COL_DESCRIPTION_COURSE],
+                'pictureId'   =>       $row[StudyPressDB::COL_PICTURE_COURSE],
                 'avancement'  =>       $row[StudyPressDB::COL_AVANCEMENT_COURSE],
                 'postId'      => (int) $row[StudyPressDB::COL_ID_POST_COURSE]
             ))
         );
     }
 
-
+    
     public function getById($id)
     {
         $result = $this->_access->getRow($this->_access->prepare("SELECT * FROM " . StudyPressDB::getTableNameCourse() . " WHERE " . StudyPressDB::COL_ID_COURSE . " = '%d'", $id));
@@ -297,14 +358,22 @@ class CourseManager
 
         $course =  self::returnedCourse($result);
 
+        if($course) {
+
+            $course->setNbreLessons($this->getNumberOf('lesson',$course->getId()));
+            $course->setNbrequizs($this->getNumberOf('quiz',$course->getId()));
+
+            //Ajouter les categories
+            $course->setCategories($this->getCategoriesId($course->getId()));
 
 
+            //Ajouter les les groupe buddypress
+            $course->setGroupsBP($this->getGroupsBPId($course->getId()));
 
 
-        $course->setCategories($this->getCategoriesId($course->getId()));
-
-
-        $course->setAuthors($this->getUsersId($course->getId()));
+            //Ajouter les users
+            $course->setAuthors($this->getUsersId($course->getId()));
+        }
 
 
         return $course;
@@ -314,8 +383,13 @@ class CourseManager
     public function getCoursesByAuthor($authorId)
     {
 
+        $user = new StudyPressUserWP($authorId);
+        if($user->isAdministrator())
+        {
+            return $this->getAll();
+        }
         $courses = array();
-        $result = $this->_access->getResults($this->_access->prepare("SELECT ". StudyPressDB::COL_ID_COURSE_USERS_N_COURSE." FROM " .  StudyPressDB::getTableName_CourseUsers()." WHERE " . StudyPressDB::COL_ID_USERS_USERS_N_COURSE ." = '%d'",$authorId));
+        $result = $this->_access->getResults($this->_access->prepare("SELECT ". StudyPressDB::COL_ID_COURSE_USERS_N_COURSE." FROM " .  StudyPressDB::getTableName_CourseUsers()." WHERE " . StudyPressDB::COL_ID_USERS_USERS_N_COURSE ." = '%d' ORDER BY " . StudyPressDB::COL_ID_COURSE_USERS_N_COURSE ." DESC ",$authorId));
 
         foreach ($result as $row) {
 
@@ -357,16 +431,20 @@ class CourseManager
 
 
         $post = array(
-            'post_content' => '',
-            'post_name' => $course->getName(),
-            'post_title' => $course->getName(),
-            'post_status' => 'publish',
-            'post_type' => 'course',
-            'post_category' => $course->getCategories()
+            // 'ID'             => [ <post id> ] // Are you updating an existing post?
+            'post_content' => '',  // The full text of the post.
+            'post_name' => $course->getName(),// The name (slug) for your post
+            'post_title' => $course->getName(), // The title of your post.
+            'post_status' => 'publish',// [ 'draft' | 'publish' | 'pending'| 'future' | 'private' | custom registered status ] // Default 'draft'.
+            'post_type' => 'course',//[ 'post' | 'page' | 'link' | 'nav_menu_item' | custom post type ] // Default 'post'.
+            //'post_author' => $lesson->getAuthorId(),// The user ID number of the author. Default is the current user ID.
+
+            'post_category' => $course->getCategories(),// Default empty.
+            'post_excerpt' => ($course->getDescription())?$course->getDescription():$course->getName()
 
         );
 
-
+        
 
 
 
@@ -375,24 +453,40 @@ class CourseManager
 
         $post = array(
             'ID'           => $post_id,
-            'post_content' => "[studypress_child id=". $post_id ."]"
+            'post_content' => "[studypress_child id=". $course->getId() ."]"
         );
 
+
         PostWP::updatePost( $post );
-
-
-    
-
-
-
-
-
-
-
 
         $course->setPostId($post_id);
 
         $this->update($course->getId(), $course);
+
+        
+    }
+
+
+    public function updatePost(Course $course){
+
+        $post = array(
+            'ID'             => $course->getPostId(),// Are you updating an existing post?
+            'post_content' => "[studypress_child id=". $course->getId() ."]",
+            'post_name' => $course->getName(),// The name (slug) for your post
+            'post_title' => $course->getName(), // The title of your post.
+            'post_status' => 'publish',// [ 'draft' | 'publish' | 'pending'| 'future' | 'private' | custom registered status ] // Default 'draft'.
+            'post_type' => 'course',//[ 'post' | 'page' | 'link' | 'nav_menu_item' | custom post type ] // Default 'post'.
+            //'post_author' => $lesson->getAuthorId(),// The user ID number of the author. Default is the current user ID.
+
+            'post_category' => $course->getCategories(),// Default empty.
+
+            'post_excerpt' => ($course->getDescription())?$course->getDescription():$course->getName()
+
+        );
+
+        PostWP::updatePost($post);
+
+
 
     }
 
@@ -400,6 +494,48 @@ class CourseManager
     public function unpost(Course $course){
         PostWP::unPost( $course->getPostId() );
 
+
+    }
+
+
+    public function attacherImageToPost(Course $course)
+    {
+        if($course->getPictureId() && $course->getPostId()){
+
+            PostWP::setPostPicture(  $course->getPostId(), $course->getPictureId() );
+        }
+    }
+
+
+
+    public function dettacherImageFromPost(Course $course)
+    {
+        if(PostWP::hasPostPicture($course->getId()))
+        {
+            PostWP::deletePostPicture( $course->getId() );
+        }
+    }
+
+
+    public function getActivitiesOfCourse($courseId){
+
+        $activities = array();
+        $result = $this->_access->getResults($this->_access->prepare("SELECT * FROM " .  StudyPressDB::getTableNameActivity() . " WHERE " . StudyPressDB::COL_ID_COURSE_ACTIVITY . " = '%d' ORDER BY " . StudyPressDB::COL_ORDER_ACTIVITY ." ASC",$courseId));
+
+
+        foreach ($result as $row) {
+            $activity = array();
+
+            if($row[StudyPressDB::COL_TYPE_ACTIVITY] === "lesson")
+                $activity = LessonManager::returnedLesson($row);
+            else
+                $activity = QuizManager::returnedQuiz($row);
+
+            if($activity) $activities[] = $activity;
+        }
+
+
+        return $activities;
 
     }
